Packaging the code for future deployment
New dependencies are required for running the code, especially it relies on a C++ parallel version of dbscan. Also xarray are used to read netcdf files (specifically the paleosurface dataset and output from the dbscan algorithm).
The package also uses a YAML input file for defining the initial conditions and part of the inputs required the position of each plate ID over time (this can be extracted from gplates directly).
Note that at the moment, the package does not contain the code regarding seafloor spreading.